modularized untriage assignment
